move Shape Interface to inside package
