+++ src/cs3500/ExCELlence/model/AnimationModel.java
@@ -1,5 +1,6 @@
 package cs3500.ExCELlence.model;
 
+import cs3500.ExCELlence.model.shapes.Shape;
 import cs3500.ExCELlence.model.transitions.Transition;import java.util.Stack;
 
 /**
