+++ src/cs3500/ExCELlence/model/shapes/Oval.java
@@ -1,9 +1,8 @@
 package cs3500.ExCELlence.model.shapes;
 
-import cs3500.ExCELlence.model.Shape;
-
 public class Oval extends ShapeImpl implements Shape {
   public void draw() {
+    // Implementation-dependent of oval shape, not yet clear of how to draw the shape
 
   }
 }
